+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Update Uppy.js to v3.0.1
 - Resolve nested linked content items on content item population
 - Auto-populate content item for content previews
+- Fix response cache
 
 ## 2.3.0 (2022-09-16)